requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

tensorflow==2.14.0
tensorflow-hub==0.15.0
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.9.0

redis==5.0.1
requests==2.31.0
httpx==0.25.2

pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

black==23.12.0
flake8==6.1.0
pylint==3.0.3
